+++ Frontend/tkd_hub_web/src/app/components/contact/contactLinks.ts
@@ -1,0 +1,29 @@
+const contactLinks = [
+  {
+    href: "https://wa.me/1234567890",
+    icon: "bi bi-whatsapp",
+    title: "WhatsApp",
+    className: "text-green-500 hover:text-green-600 transition",
+  },
+  {
+    href: "https://instagram.com/yourprofile",
+    icon: "bi bi-instagram",
+    title: "Instagram",
+    className: "text-pink-500 hover:text-pink-600 transition",
+  },
+  {
+    href: "https://facebook.com/yourprofile",
+    icon: "bi bi-facebook",
+    title: "Facebook",
+    className: "text-blue-700 hover:text-blue-800 transition",
+  },
+  {
+    href: "https://youtube.com/yourchannel",
+    icon: "bi bi-youtube",
+    title: "YouTube",
+    className: "text-red-600 hover:text-red-700 transition",
+  },
+];
+
+
+export default contactLinks;
