+++ src/TKDHubAPI.Application/DTOs/Promotion/PromotionDto.cs
@@ -7,9 +7,9 @@ public class PromotionDto
 {
     public int Id { get; set; }
     public int StudentId { get; set; }
-    public string? StudentName { get; set; }      // Added
+    public string? StudentName { get; set; }
     public int RankId { get; set; }
-    public string? RankName { get; set; }         // Added
+    public string? RankName { get; set; }
     public DateTime PromotionDate { get; set; }
     public int CoachId { get; set; }
     public string? Notes { get; set; }
