+++ src/TKDHubAPI.Application/DTOs/User/CreateUserDto.cs
@@ -12,5 +12,5 @@ public class CreateUserDto
     public int? DojaangId { get; set; }
     public int? RankId { get; set; } // Renamed from CurrentRankId to RankId
     public DateTime? JoinDate { get; set; }
-    public List<int> RoleIds { get; set; } // Add this if you want to assign roles on creation
+    public List<int> RoleIds { get; set; } // Assign roles on creation
 }
